feat(asyncengine): refactor recording/submission/presentation
- Refactor: remove simulator-owned per-surface helpers and delegate
  command recording, submission, and presentation responsibilities to
  the render-graph / renderer integration. The simulator now exposes a
  per-frame `FrameContext::presentable_flags` for worker-to-frame
  readiness signaling and forwards presentation to the graphics layer
  via `GraphicsLayer::PresentSurfaces(...)`.

- Implementation: added `GraphicsLayer::PresentSurfaces(const
  std::vector<RenderSurface>&)` (simulated present + logging), updated
  `PhasePresent` to gather presentable surfaces and call the new API,
  and removed simulator helpers that performed per-surface
  record/submit/present.

- Rationale: aligns ownership with the renderer, avoids mutating
  frame-owned surface state across threads, and provides a minimal, safe
  atomic-based signaling mechanism for worker threads.

- Follow-ups / TODOs:
  1. Producer-side flag setting: have the render-graph executor or
     renderer modules perform
     `std::atomic_ref<uint8_t>(flags[i]).store(1,
     std::memory_order_release)` when a view is ready.
  2. Replace the demo copy-path with non-mutable references or
     `std::span` to avoid copying `RenderSurface` values at present
     time.
  3. Consider using an explicit atomic array or compact bitset for
     `presentable_flags` for efficiency and clarity.

+++ projects/Oxygen.Engine/Examples/AsyncEngine/Examples/GameModule.h
@@ -178,7 +178,7 @@ private:
   // Resource handles
   uint32_t player_entity_handle_ { 0 };
   uint32_t world_state_handle_ { 0 };
-  std::vector<uint32_t> dynamic_entities_;
+  std::vector<uint64_t> dynamic_entities_;
 
   // Statistics
   uint32_t input_events_processed_ { 0 };
